I think I have everything working now the way I want it to. The translation and scaling have all been adjusted properly when the slice occurs, so no matter how you move the image, the intended area is always sliced properly.
Personal note:
Progress is incredibly slow and many small mistakes and misunderstandings that slow this project down are due to having to balance my time with a newborn baby. I am lucky if I get even 3 minutes to focus on a problem which is not enough to make even the slightest change to the code or to begin to grasp a problem.